// src/app/routing/enums/AuthRoute.enum.ts
export enum AuthRoute {
  dashboard = '/',
  questions = '/questions',
  oshcertificate = '/ohs-certificate-requests',
  performanceCardDetail = '/performance-card-details/:id',
  PreviewPerformanceCardCertificate = '/pcrequest/performance-certificate/:id',
  EstablishmentList = '/establishment-list',
  VistListRequest = '/visits',
  startSelfAssessment = '/start-self-assessment',
  SelfAssessmentResult = '/self-assessment-result',
}